Make sentence more precise
another option (which I would actually prefer) is to completely remove the sentence. Why then desktop wallets (https://docs.ethhub.io/using-ethereum/wallets/desktop/) do not have a sentence like this as a con? Often the security of e.g. windows is weaker than a Android device as it is not isolating/sandboxing applications the same way.
But if the sentence does not get removed it should at least be precise. You can use mobile wallets like WallETH in combination with a hardware wallet (via USB-OTG)